fix how we find parent process
